:tada: feat: able to add notes to vehicles
